+++ utilities/libiconv/lib/uhc_1.h
@@ -1,21 +1,20 @@
 /*
- * Copyright (C) 1999-2001 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2012, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
- * and/or modify it under the terms of the GNU Library General Public
- * License as published by the Free Software Foundation; either version 2
+ * and/or modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either version 2.1
  * of the License, or (at your option) any later version.
  *
  * The GNU LIBICONV Library is distributed in the hope that it will be
  * useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Library General Public License for more details.
+ * Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU Library General Public
+ * You should have received a copy of the GNU Lesser General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
- * Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <https://www.gnu.org/licenses/>.
  */
 
 /*
@@ -804,7 +803,7 @@ static const unsigned char uhc_1_2uni_page81[5696] = {
 };
 
 static int
-uhc_1_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+uhc_1_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   unsigned char c1 = s[0];
   if ((c1 >= 0x81 && c1 <= 0xa0)) {
@@ -1697,7 +1696,7 @@ static const Summary16 uhc_1_uni2indx_pageac[459] = {
 };
 
 static int
-uhc_1_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+uhc_1_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (n >= 2) {
     if (wc >= 0xac00 && wc < 0xc8b0) {
@@ -1706,9 +1705,9 @@ uhc_1_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
       unsigned int i = wc & 0x0f;
       if (used & ((unsigned short) 1 << i)) {
         unsigned short c;
-        /* Keep in `used' only the bits 0..i-1. */
+        /* Keep in 'used' only the bits 0..i-1. */
         used &= ((unsigned short) 1 << i) - 1;
-        /* Add `summary->indx' and the number of bits set in `used'. */
+        /* Add 'summary->indx' and the number of bits set in 'used'. */
         used = (used & 0x5555) + ((used & 0xaaaa) >> 1);
         used = (used & 0x3333) + ((used & 0xcccc) >> 2);
         used = (used & 0x0f0f) + ((used & 0xf0f0) >> 4);
