+++ utilities/libiconv/lib/utf16le.h
@@ -1,21 +1,20 @@
 /*
- * Copyright (C) 1999-2001, 2008 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2008, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
- * and/or modify it under the terms of the GNU Library General Public
- * License as published by the Free Software Foundation; either version 2
+ * and/or modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either version 2.1
  * of the License, or (at your option) any later version.
  *
  * The GNU LIBICONV Library is distributed in the hope that it will be
  * useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Library General Public License for more details.
+ * Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU Library General Public
+ * You should have received a copy of the GNU Lesser General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street,
- * Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <https://www.gnu.org/licenses/>.
  */
 
 /*
@@ -25,7 +24,7 @@
 /* Specification: RFC 2781 */
 
 static int
-utf16le_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+utf16le_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   int count = 0;
   if (n >= 2) {
@@ -52,7 +51,7 @@ ilseq:
 }
 
 static int
-utf16le_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+utf16le_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   if (!(wc >= 0xd800 && wc < 0xe000)) {
     if (wc < 0x10000) {
